Make it more clear that multiple Elm files can be formatted as a batch
The CLI help docs made it seem like elm-format could only be invoked
against a single Elm source file, but the README includes an example
of how elm-format can be used to recursively format all .elm files
in a directory. This change brings the CLI help docs in-line with the
README and the capabilities of the tool.